created new main layout for the whole app
