+++ src/layout/BasicLayout/index.ts
@@ -1,0 +1,1 @@
+export { BasicLayout } from './BasicLayout';
