+++ src/navigation/routes.tsx
@@ -7,47 +7,53 @@ import VulnerabilityForm from 'components/Vulnerability/VulnerabilityForm';
 import CommitteeTools from 'components/CommitteeTools/CommitteTools';
 import VaultEditor from 'components/VaultEditor/VaultEditor';
 import AirdropMachine from 'components/AirdropMachine/AirdropMachine';
+import { BasicLayout } from 'layout';
 
 const routes: RouteObject[] = [
   {
-    path: '/',
-    element: <Navigate to={RoutePaths.vaults} replace={true} />,
-  },
-  {
-    path: `${RoutePaths.vaults}`,
-    element: <Honeypots />,
-  },
-  {
-    path: `${RoutePaths.vaults}/:pid`,
-    element: <Honeypots />,
-  },
-  {
-    path: `${RoutePaths.vaults}/:pid/deposit`,
-    element: <Honeypots showDeposit={true} />,
-  },
-  {
-    path: `${RoutePaths.gov}`,
-    element: <Gov />,
-  },
-  {
-    path: `${RoutePaths.vulnerability}`,
-    element: <VulnerabilityForm />,
-  },
-  {
-    path: `${RoutePaths.committee_tools}`,
-    element: <CommitteeTools />,
-  },
-  {
-    path: `${RoutePaths.vault_editor}`,
-    element: <VaultEditor />,
-  },
-  {
-    path: `${RoutePaths.vault_editor}/:ipfsHash`,
-    element: <VaultEditor />,
-  },
-  {
-    path: `${RoutePaths.airdrop_machine}`,
-    element: <AirdropMachine />,
+    element: <BasicLayout />,
+    children: [
+      {
+        path: '/',
+        element: <Navigate to={RoutePaths.vaults} replace={true} />,
+      },
+      {
+        path: `${RoutePaths.vaults}`,
+        element: <Honeypots />,
+      },
+      {
+        path: `${RoutePaths.vaults}/:pid`,
+        element: <Honeypots />,
+      },
+      {
+        path: `${RoutePaths.vaults}/:pid/deposit`,
+        element: <Honeypots showDeposit={true} />,
+      },
+      {
+        path: `${RoutePaths.gov}`,
+        element: <Gov />,
+      },
+      {
+        path: `${RoutePaths.vulnerability}`,
+        element: <VulnerabilityForm />,
+      },
+      {
+        path: `${RoutePaths.committee_tools}`,
+        element: <CommitteeTools />,
+      },
+      {
+        path: `${RoutePaths.vault_editor}`,
+        element: <VaultEditor />,
+      },
+      {
+        path: `${RoutePaths.vault_editor}/:ipfsHash`,
+        element: <VaultEditor />,
+      },
+      {
+        path: `${RoutePaths.airdrop_machine}`,
+        element: <AirdropMachine />,
+      },
+    ]
   },
   {
     path: '*',
